ECHS_End_Sem/requirements.txt
flask
opencv-python-headless
torch
ultralytics
numpy
Pillow
